Update _get_account_id to ignore non-brokerage accounts.
Non-brokerage accounts will cause a timeout and crash the bot. This fixes that by checking for "brokerage" in the name, which matches the list of trade-able accounts in the order-page dialog.

However, this will also not show holdings for these accounts, which is probably appropriate given that they are edge-cases like cash-plus.